% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.calculate_class_weights}
\alias{.calculate_class_weights}
\title{Calculate Class Weights}
\usage{
.calculate_class_weights(y, class_levels)
}
\arguments{
\item{y}{Factor or character target}

\item{class_levels}{Character vector of class levels}
}
\value{
Named list of class weights
}
\description{
Calculate Class Weights
}
\keyword{internal}
