% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_package_version}
\alias{.check_package_version}
\title{Check Package Version}
\usage{
.check_package_version(pkg, min_version)
}
\arguments{
\item{pkg}{Package name}

\item{min_version}{Minimum required version}
}
\description{
Check Package Version
}
\keyword{internal}
