% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.create_rolling_features}
\alias{.create_rolling_features}
\title{Create Rolling Features}
\usage{
.create_rolling_features(data, features, rolling_config)
}
\description{
Create Rolling Features
}
\keyword{internal}
