% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cross_validate}
\alias{.cross_validate}
\title{Cross-Validation for easyNNR}
\usage{
.cross_validate(data, target, k = 5, ...)
}
\arguments{
\item{data}{Data frame}

\item{target}{Target column name}

\item{k}{Number of folds}

\item{...}{Additional arguments passed to easy_nn}
}
\value{
List with CV results
}
\description{
Cross-Validation for easyNNR
}
\keyword{internal}
