% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.extract_date_features}
\alias{.extract_date_features}
\title{Extract Features from Date Columns}
\usage{
.extract_date_features(data, date_cols)
}
\description{
Extract Features from Date Columns
}
\keyword{internal}
