% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fmt}
\alias{.fmt}
\title{Format Number for Printing}
\usage{
.fmt(x, digits = 4)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places (default: 4)}
}
\description{
Format Number for Printing
}
\keyword{internal}
