% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.frequency_encode}
\alias{.frequency_encode}
\title{Frequency Encoding for Categorical Variables}
\usage{
.frequency_encode(data, cat_cols)
}
\description{
Frequency Encoding for Categorical Variables
}
\keyword{internal}
