% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.infer_task}
\alias{.infer_task}
\title{Infer Task Type from Target Variable}
\usage{
.infer_task(y)
}
\arguments{
\item{y}{Target variable vector}
}
\description{
Infer Task Type from Target Variable
}
\keyword{internal}
