% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.lasso_selection}
\alias{.lasso_selection}
\title{L1/LASSO-based Feature Selection}
\usage{
.lasso_selection(X, y, n_features, task)
}
\description{
L1/LASSO-based Feature Selection
}
\keyword{internal}
