% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_optimizer}
\alias{.make_optimizer}
\title{Make Keras Optimizer from String or Object}
\usage{
.make_optimizer(opt, lr = 0.001)
}
\arguments{
\item{opt}{Character string or keras optimizer object}

\item{lr}{Learning rate (default: 0.001)}
}
\description{
Make Keras Optimizer from String or Object
}
\keyword{internal}
