% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.mutual_info_scores}
\alias{.mutual_info_scores}
\title{Calculate Mutual Information Scores}
\usage{
.mutual_info_scores(X, y, task)
}
\description{
Calculate Mutual Information Scores
}
\keyword{internal}
