% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.permutation_importance}
\alias{.permutation_importance}
\title{Calculate Feature Importance from Neural Network}
\usage{
.permutation_importance(model, X_test, y_test, feature_names = NULL)
}
\arguments{
\item{model}{Keras model}

\item{X_test}{Test feature matrix}

\item{y_test}{Test target vector}

\item{feature_names}{Character vector of feature names}
}
\value{
Data frame with feature importance scores
}
\description{
Calculate Feature Importance from Neural Network
}
\keyword{internal}
