% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.rfe_selection}
\alias{.rfe_selection}
\title{Recursive Feature Elimination (Simplified)}
\usage{
.rfe_selection(X, y, n_features, task)
}
\description{
Recursive Feature Elimination (Simplified)
}
\keyword{internal}
