% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.target_encode}
\alias{.target_encode}
\title{Target Encoding for Categorical Variables}
\usage{
.target_encode(data, cat_cols, target, smoothing = 1)
}
\description{
Target Encoding for Categorical Variables
}
\keyword{internal}
