% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.yeojohnson_inverse}
\alias{.yeojohnson_inverse}
\title{Yeo-Johnson Inverse Transformation}
\usage{
.yeojohnson_inverse(y, lambda)
}
\description{
Yeo-Johnson Inverse Transformation
}
\keyword{internal}
