% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{easy_compare}
\alias{easy_compare}
\title{Compare Multiple easyNNR Models}
\usage{
easy_compare(..., names = NULL)
}
\arguments{
\item{...}{easyNNR model objects to compare}

\item{names}{Optional character vector of model names}
}
\value{
A tibble comparing model configurations and performance
}
\description{
Create a comparison table of multiple trained models.
}
\examples{
\dontrun{
model1 <- easy_nn(iris, target = "Species", layers = c(64, 32))
model2 <- easy_nn(iris, target = "Species", layers = c(128, 64, 32))
comparison <- easy_compare(model1, model2, names = c("Small", "Large"))
}

}
