% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{easy_dashboard}
\alias{easy_dashboard}
\title{Create a Dashboard of Plots}
\usage{
easy_dashboard(object)
}
\arguments{
\item{object}{An easyNNR model}
}
\value{
A combined ggplot2 object (uses patchwork if available)
}
\description{
Generate a multi-panel dashboard with key visualizations.
}
