% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{easy_plot_class_distribution}
\alias{easy_plot_class_distribution}
\title{Plot Class Distribution (Classification Only)}
\usage{
easy_plot_class_distribution(object, title = "Class Distribution")
}
\arguments{
\item{object}{An easyNNR classification model}

\item{title}{Plot title}
}
\value{
A ggplot2 object
}
\description{
Create a bar plot showing the distribution of classes in training data
and predictions.
}
