% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{easy_plot_residual_dist}
\alias{easy_plot_residual_dist}
\title{Plot Residual Distribution (Regression Only)}
\usage{
easy_plot_residual_dist(object, title = "Residual Distribution")
}
\arguments{
\item{object}{An easyNNR regression model}

\item{title}{Plot title}
}
\value{
A ggplot2 object
}
\description{
Create a histogram and density plot of residuals for regression models.
}
