% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{end_training_session}
\alias{end_training_session}
\title{End Training Session}
\usage{
end_training_session()
}
\description{
End a training session and close the log file.
}
\examples{
\dontrun{
start_training_session("~/my_logs")
model <- easy_nn(iris, target = "Species")
end_training_session()
}
}
