% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.easyNNR}
\alias{predict.easyNNR}
\title{Predict Method for easyNNR Objects}
\usage{
\method{predict}{easyNNR}(object, newdata, type = NULL, ...)
}
\arguments{
\item{object}{An easyNNR model object}

\item{newdata}{New data to make predictions on}

\item{type}{Type of prediction ("class", "prob", or "response")}

\item{...}{Additional arguments (currently unused)}
}
\value{
Predictions (format depends on type)
}
\description{
S3 method for predict() generic. Calls easy_predict() internally.
}
