% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.easyNNR}
\alias{print.easyNNR}
\title{Print Method for easyNNR Objects}
\usage{
\method{print}{easyNNR}(x, ...)
}
\arguments{
\item{x}{An easyNNR object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for easyNNR Objects
}
