% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{save_easyNNR_model}
\alias{save_easyNNR_model}
\title{Save easyNNR Model and Results (architecture + weights + preprocessing)}
\usage{
save_easyNNR_model(
  model,
  output_dir,
  save_plots = TRUE,
  save_logs = TRUE,
  save_best = TRUE
)
}
\description{
Save easyNNR Model and Results (architecture + weights + preprocessing)
}
