% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{start_easyNNR_log}
\alias{start_easyNNR_log}
\title{Start easyNNR Logging}
\usage{
start_easyNNR_log(log_dir)
}
\arguments{
\item{log_dir}{Directory where log files should be saved.}
}
\description{
Starts logging all console output to a timestamped file.
}
