% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model.R
\name{start_training_session}
\alias{start_training_session}
\title{Save Training Session}
\usage{
start_training_session(log_dir)
}
\arguments{
\item{log_dir}{Directory to save log files}
}
\value{
Connection object (for internal use)
}
\description{
Start a training session that logs all output to a file.
}
\examples{
\dontrun{
start_training_session("~/my_logs")
model <- easy_nn(iris, target = "Species")
end_training_session()
}
}
