% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.easyNNR}
\alias{summary.easyNNR}
\title{Summary Method for easyNNR Objects}
\usage{
\method{summary}{easyNNR}(object, ...)
}
\arguments{
\item{object}{An easyNNR object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for easyNNR Objects
}
